@@ -3,67 +3,66 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using Microsoft.CodeAnalysis;
 
-namespace NForza.Cyrus.Generators.WebApi
+namespace NForza.Cyrus.Generators.WebApi;
+
+internal static class TypeAnnotations
 {
-    internal static class TypeAnnotations
+    public static string AugmentRouteWithTypeAnnotations(string path, ITypeSymbol typeSymbol)
     {
-        public static string AugmentRouteWithTypeAnnotations(string path, ITypeSymbol typeSymbol)
+        return Regex.Replace(path, @"\{(?<param>\w+)\}", match =>
         {
-            return Regex.Replace(path, @"\{(?<param>\w+)\}", match =>
-            {
-                var paramName = match.Groups["param"].Value;
+            var paramName = match.Groups["param"].Value;
 
-                var property = typeSymbol
-                    .GetMembers()
-                    .OfType<IPropertySymbol>()
-                    .FirstOrDefault(p => p.Name.Equals(paramName, StringComparison.OrdinalIgnoreCase));
+            var property = typeSymbol
+                .GetMembers()
+                .OfType<IPropertySymbol>()
+                .FirstOrDefault(p => p.Name.Equals(paramName, StringComparison.OrdinalIgnoreCase));
 
-                if (property is null)
-                    return match.Value;
+            if (property is null)
+                return match.Value;
 
-                var constraint = GetRouteConstraint(property.Type);
-                return constraint is null
-                    ? match.Value
-                    : $"{{{paramName}:{constraint}}}";
-            });
-        }
+            var constraint = GetRouteConstraint(property.Type);
+            return constraint is null
+                ? match.Value
+                : $"{{{paramName}:{constraint}}}";
+        });
+    }
 
-        static string? GetRouteConstraint(ITypeSymbol symbol)
+    static string? GetRouteConstraint(ITypeSymbol symbol)
+    {
+        if (symbol is INamedTypeSymbol namedType)
         {
-            if (symbol is INamedTypeSymbol namedType)
+            if (namedType.OriginalDefinition.SpecialType == SpecialType.System_Nullable_T
+                && namedType.TypeArguments.Length == 1)
             {
-                if (namedType.OriginalDefinition.SpecialType == SpecialType.System_Nullable_T
-                    && namedType.TypeArguments.Length == 1)
-                {
-                    return GetRouteConstraint(namedType.TypeArguments[0]);
-                }
-
-                var attributes = namedType.GetAttributes();
-
-                if (attributes.Any(a => a.AttributeClass?.Name == "GuidIdAttribute"))
-                    return "guid";
-
-                if (attributes.Any(a => a.AttributeClass?.Name == "IntIdAttribute"))
-                    return "int";
-
-                if (attributes.Any(a => a.AttributeClass?.Name == "StringIdAttribute"))
-                    return null;
-
-                var typeName = namedType.ToDisplayString();
-
-                return typeName switch
-                {
-                    "System.Guid" => "guid",
-                    "System.Int32" => "int",
-                    "System.Int64" => "long",
-                    "System.Boolean" => "bool",
-                    "System.Single" => "float",
-                    "System.Double" => "double",
-                    _ => null
-                };
+                return GetRouteConstraint(namedType.TypeArguments[0]);
             }
 
-            return null;
+            var attributes = namedType.GetAttributes();
+
+            if (attributes.Any(a => a.AttributeClass?.Name == "GuidIdAttribute"))
+                return "guid";
+
+            if (attributes.Any(a => a.AttributeClass?.Name == "IntIdAttribute"))
+                return "int";
+
+            if (attributes.Any(a => a.AttributeClass?.Name == "StringIdAttribute"))
+                return null;
+
+            var typeName = namedType.ToDisplayString();
+
+            return typeName switch
+            {
+                "System.Guid" => "guid",
+                "System.Int32" => "int",
+                "System.Int64" => "long",
+                "System.Boolean" => "bool",
+                "System.Single" => "float",
+                "System.Double" => "double",
+                _ => null
+            };
         }
+
+        return null;
     }
 }
