@@ -2,50 +2,49 @@
 using System.Reflection;
 using Fluid;
 
-namespace NForza.Cyrus.Templating
+namespace NForza.Cyrus.Templating;
+
+public class LiquidEngine
 {
-    public class LiquidEngine
+    TemplateResourceFileProvider fileProvider;
+    private TemplateOptions options;
+
+    public LiquidEngine(Assembly assembly, Action<TemplateOptions> configure = null)
     {
-        TemplateResourceFileProvider fileProvider;
-        private TemplateOptions options;
+        fileProvider = new TemplateResourceFileProvider(assembly);
+        options = new TemplateOptions();
+        options.Filters.AddFilter("as-contract", CyrusFilters.ContractName);
+        options.Filters.AddFilter("generated-hub-name", CyrusFilters.GeneratedHubName);
+        options.Filters.AddFilter("camel-cased", CyrusFilters.CamelCased);
+        options.FileProvider = fileProvider;
+        options.MemberAccessStrategy = UnsafeMemberAccessStrategy.Instance;
+        configure?.Invoke(options);
+    }
 
-        public LiquidEngine(Assembly assembly, Action<TemplateOptions> configure = null)
+    private TemplateContext GetContext(object model)
+    {
+        return new TemplateContext(model, options);
+    }
+
+    private IFluidTemplate GetTemplate(string path)
+    {
+        var parser = new FluidParser();
+        var template = parser.Parse(fileProvider.GetTemplateContents(path));
+        return template;
+    }
+
+    public string Render(object model, string templateName)
+    {
+        try
         {
-            fileProvider = new TemplateResourceFileProvider(assembly);
-            options = new TemplateOptions();
-            options.Filters.AddFilter("as-contract", CyrusFilters.ContractName);
-            options.Filters.AddFilter("generated_hub_name", CyrusFilters.GeneratedHubName);
-            options.Filters.AddFilter("camel-cased", CyrusFilters.CamelCased);
-            options.FileProvider = fileProvider;
-            options.MemberAccessStrategy = UnsafeMemberAccessStrategy.Instance;
-            configure?.Invoke(options);
+            IFluidTemplate template = GetTemplate(templateName);
+            var result = template.Render(GetContext(model));
+            return result;
+
         }
-
-        private TemplateContext GetContext(object model)
+        catch (ParseException e)
         {
-            return new TemplateContext(model, options);
-        }
-
-        private IFluidTemplate GetTemplate(string path)
-        {
-            var parser = new FluidParser();
-            var template = parser.Parse(fileProvider.GetTemplateContents(path));
-            return template;
-        }
-
-        public string Render(object model, string templateName)
-        {
-            try
-            {
-                IFluidTemplate template = GetTemplate(templateName);
-                var result = template.Render(GetContext(model));
-                return result;
-
-            }
-            catch (ParseException e)
-            {
-                throw new InvalidOperationException($"Can't load template \"{templateName}\": {e.Message}", e);
-            }
+            throw new InvalidOperationException($"Can't load template \"{templateName}\": {e.Message}", e);
         }
     }
 }
