@@ -6,7 +6,7 @@ using NForza.Cyrus.Generators.Model;
 using NForza.Cyrus.Generators.Roslyn;
 using NForza.Cyrus.Templating;
 
-namespace NForza.Cyrus.Generators.Cqrs;
+namespace NForza.Cyrus.Generators.Events;
 
 public class EventGenerator : CyrusGeneratorBase
 {
