@@ -5,7 +5,7 @@ using Microsoft.CodeAnalysis.Text;
 using NForza.Cyrus.Generators.Roslyn;
 using NForza.Cyrus.Templating;
 
-namespace NForza.Cyrus.Generators.Generators.TypedIds;
+namespace NForza.Cyrus.Generators.TypedIds;
 
 public class StringIdGenerator : CyrusGeneratorBase
 {
